Split gallery classes and add more explanation to validation example
